Added NetworkRequest flow
The new flow simulates a request originating from the UI (main thread)
but made from a separate thread that awaits the answer from the server
without blocking the main thread. The flow also simulates the server work
with a third thread.
A coroutine, Generator and FutureAwaiter pattern is used to co_await the server.